requirements.txt
Django==3.1.1
djangorestframework==3.11.1
ipython==7.18.1
django-extensions==3.0.8
uuid==1.30
psycopg2_binary==2.8.6
requests==2.24.0
celery==4.4.7
redis==3.5.3
django-redis==4.12.1
dj-database-url==0.5.0
django-enum-choices==2.1.2
